rename fragments, simplify view for users
