+++ examples/react-navigation/android-fragment/android/app/src/main/java/com/example/bottomnavigationviews/MainActivity.kt
@@ -7,8 +7,8 @@ import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import androidx.viewpager2.widget.ViewPager2
 import com.example.bottomnavigationviews.databinding.ActivityMainBinding
-import com.example.bottomnavigationviews.ui.dashboard.DashboardFragment
 import com.example.bottomnavigationviews.ui.home.HomeFragment
+import com.example.bottomnavigationviews.ui.react_native.ReactNativeFragment
 import com.facebook.react.ReactFragment
 import com.facebook.react.modules.core.DefaultHardwareBackBtnHandler
 import com.google.android.material.bottomnavigation.BottomNavigationView
@@ -31,7 +31,7 @@ class MainActivity : AppCompatActivity(), DefaultHardwareBackBtnHandler {
         pager.adapter = object : FragmentStateAdapter(this) {
             private val fragments = arrayOf(
                 HomeFragment(),
-                DashboardFragment(),
+                ReactNativeFragment(),
                 // Put the ReactNative Fragment here!
                 ReactFragment.Builder()
                     .setComponentName("BugsnagExampleApp")
@@ -46,8 +46,7 @@ class MainActivity : AppCompatActivity(), DefaultHardwareBackBtnHandler {
         navView.setOnItemSelectedListener { item ->
             when (item.itemId) {
                 R.id.navigation_home -> pager.currentItem = 0
-                R.id.navigation_dashboard -> pager.currentItem = 1
-                R.id.navigation_notifications -> pager.currentItem = 2
+                R.id.navigation_react_native -> pager.currentItem = 2
             }
 
             true
