+++ examples/react-navigation/android-fragment/android/app/src/main/java/com/example/bottomnavigationviews/ui/home/HomeViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.ViewModel
 class HomeViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "Use the navigation to navigate to the React Native app"
     }
     val text: LiveData<String> = _text
 }
