+++ examples/react-navigation/android-fragment/android/app/src/main/java/com/example/bottomnavigationviews/ui/react_native/ReactNativeFragment.kt
@@ -1,4 +1,4 @@
-package com.example.bottomnavigationviews.ui.dashboard
+package com.example.bottomnavigationviews.ui.react_native
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.example.bottomnavigationviews.databinding.FragmentDashboardBinding
+import com.example.bottomnavigationviews.databinding.FragmentReactNativeBinding
 
-class DashboardFragment : Fragment() {
+class ReactNativeFragment : Fragment() {
 
-    private var _binding: FragmentDashboardBinding? = null
+    private var _binding: FragmentReactNativeBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,14 +22,14 @@ class DashboardFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val dashboardViewModel =
-            ViewModelProvider(this).get(DashboardViewModel::class.java)
+        val reactNativeViewModel =
+            ViewModelProvider(this).get(ReactNativeViewModel::class.java)
 
-        _binding = FragmentDashboardBinding.inflate(inflater, container, false)
+        _binding = FragmentReactNativeBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textDashboard
-        dashboardViewModel.text.observe(viewLifecycleOwner) {
+        val textView: TextView = binding.textReactNative
+        reactNativeViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
         return root
