+++ examples/react-navigation/android-fragment/android/app/src/main/java/com/example/bottomnavigationviews/ui/react_native/ReactNativeViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.bottomnavigationviews.ui.notifications
+package com.example.bottomnavigationviews.ui.react_native
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class ReactNativeViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
+        value = "This is react native Fragment"
     }
     val text: LiveData<String> = _text
 }
