+++ examples/react-navigation/android-fragment/android/app/src/main/res/drawable/ic_react_native_black_24dp.xml
