+++ examples/react-navigation/android-fragment/android/app/src/main/res/layout/fragment_react_native.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.dashboard.DashboardFragment">
+    tools:context=".ui.react_native.ReactNativeFragment">
 
     <TextView
-        android:id="@+id/text_dashboard"
+        android:id="@+id/text_react_native"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
