+++ examples/react-navigation/android-fragment/android/app/src/main/res/menu/bottom_nav_menu.xml
@@ -7,13 +7,8 @@
         android:title="@string/title_home" />
 
     <item
-        android:id="@+id/navigation_dashboard"
-        android:icon="@drawable/ic_dashboard_black_24dp"
-        android:title="@string/title_dashboard" />
-
-    <item
-        android:id="@+id/navigation_notifications"
-        android:icon="@drawable/ic_notifications_black_24dp"
-        android:title="@string/title_notifications" />
+        android:id="@+id/navigation_react_native"
+        android:icon="@drawable/ic_react_native_black_24dp"
+        android:title="@string/title_react_native" />
 
 </menu>
