+++ examples/react-navigation/android-fragment/android/app/src/main/res/navigation/mobile_navigation.xml
@@ -12,14 +12,8 @@
         tools:layout="@layout/fragment_home" />
 
     <fragment
-        android:id="@+id/navigation_dashboard"
-        android:name="com.example.bottomnavigationviews.ui.dashboard.DashboardFragment"
-        android:label="@string/title_dashboard"
-        tools:layout="@layout/fragment_dashboard" />
-
-    <fragment
-        android:id="@+id/navigation_notifications"
-        android:name="com.example.bottomnavigationviews.ui.notifications.NotificationsFragment"
-        android:label="@string/title_notifications"
-        tools:layout="@layout/fragment_notifications" />
+        android:id="@+id/navigation_react_native"
+        android:name="com.example.bottomnavigationviews.ui.react_native.ReactNativeFragment"
+        android:label="@string/title_react_native"
+        tools:layout="@layout/fragment_react_native" />
 </navigation>
