+++ examples/react-navigation/android-fragment/android/app/src/main/res/values/strings.xml
@@ -1,6 +1,5 @@
 <resources>
     <string name="app_name">Bottom Navigation Views</string>
-    <string name="title_home">Home</string>
-    <string name="title_dashboard">Dashboard</string>
-    <string name="title_notifications">Notifications</string>
+    <string name="title_home">Android</string>
+    <string name="title_react_native">React Native</string>
 </resources>
